add task

+++ todolist/src/api/AddTask.js
@@ -1,26 +1,25 @@
-var post = require ('POST')
-
 function addTask() {
-    var check = document.getElementById('header-taskname')
+    var check = document.getElementById('taskname')
     if (check.value.trim() != '') {
-        var add = document.getElementById('header-taskname')
-        var item = document.createElement('li')
-        var currentValue = add.value.trim()
-        item.innerHTML += '<label><input type="checkbox" onclick="checkTask(event)"/>' + currentValue + '</label>'
-        item.innerHTML += '<button class="delete-button" onclick="deleteItem(event)">Delete</button>'
-        item.innerHTML += '<button class="edit-button" onclick="editTaskName(event)">Edit</button>'
-        taskList.append(item)
         var data = {
-            name : check,
+            name : check.value,
             done : false
         }
-        post ("localhost:3000/todolist", data, response => {
-            console.log(response.responseText)
+
+        postResource("http://localhost:3000/todolist", data, validate => {
+            if(validate) {
+                let taskList = document.getElementById('task-list')
+                var item = document.createElement('li')
+                var currentValue = check.value.trim()
+                item.innerHTML += '<label><input type="checkbox" onclick="checkTask(event)"/>' + currentValue + '</label>'
+                item.innerHTML += '<button class="delete-button" onclick="deleteItem(event)">Delete</button>'
+                item.innerHTML += '<button class="edit-button" onclick="editTaskName(event)">Edit</button>'
+                taskList.append(item)
+                check.value = ''
+            }
         })
     }
     else {
-        console.log('failed')
+        alert('task name should be filled')
     }
 }
-
-module.exports = addTask
